+++ staker/stakerapp.go
@@ -382,7 +382,6 @@ func (app *App) reportCriticialError(
 func (app *App) SendPhase1Transaction(
 	stakerAddr btcutil.Address,
 	stkTxHash *chainhash.Hash,
-	tag []byte,
 	covenantPks []*secp256k1.PublicKey,
 	covenantQuorum uint32,
 ) (babylonBTCDelegationTxHash string, err error) {
@@ -393,7 +392,7 @@ func (app *App) SendPhase1Transaction(
 	default:
 	}
 
-	parsedStakingTx, notifierTx, status, err := walletcontroller.StkTxV0ParsedWithBlock(app.wc, app.network, stkTxHash, tag, covenantPks, covenantQuorum)
+	parsedStakingTx, notifierTx, status, err := walletcontroller.StkTxV0ParsedWithBlock(app.wc, app.network, stkTxHash, covenantPks, covenantQuorum)
 	if err != nil {
 		app.logger.WithError(err).Info("err getting tx details")
 		return "", err