+++ CHANGELOG.md
@@ -38,6 +38,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Improvements
 
 - [#155](https://github.com/babylonlabs-io/btc-staker/pull/155) Add basic auth to daemon routes.
+- [#151](https://github.com/babylonlabs-io/btc-staker/pull/151) Load parameters from
+babylon node instead of global parameters input file in `stake-from-phase1`.
 - [#142](https://github.com/babylonlabs-io/btc-staker/pull/142) Set config default
 keyring backend to "file" type
 - [#138](https://github.com/babylonlabs-io/btc-staker/pull/138) Adds prometheus metrics
@@ -46,6 +48,8 @@ configuration to enabled it (disabled by default).
 This PR contains a series of PRs on multi-staking support and BTC staking integration.
 * [#134](https://github.com/babylonlabs-io/btc-staker/pull/134) Removal of both the watch-staking endpoint and the post-approval flow, and reduction of state in the database.
 * [#146](https://github.com/babylonlabs-io/btc-staker/pull/146) Sanity check that all transactions are standard
+* [#156](https://github.com/babylonlabs-io/btc-staker/pull/156) disable tls in config
+* [#154](https://github.com/babylonlabs-io/btc-staker/pull/154) chore: change default max feerate
 
 ## v0.15.2